+++ src/sel/random.h
@@ -22,8 +22,8 @@ namespace FT{
             
             ~Random();
            
-            vector<size_t> select(Population& pop, const MatrixXd& F, const Parameters& params);
-            vector<size_t> survive(Population& pop, const MatrixXd& F, const Parameters& params);
+            vector<size_t> select(Population& pop, const MatrixXf& F, const Parameters& params);
+            vector<size_t> survive(Population& pop, const MatrixXf& F, const Parameters& params);
             /// replaces worst individual in selected with best individual in Pop.
             void enforce_elite(Population& pop, vector<size_t>& selected);
             bool elitism;       //< whether or not to keep the best individual.
